\name{create_tree}
\alias{create_tree}
\title{
Create a phylogenetic tree from a Newick format file.
}
\description{
A function for constructing a phylogenetic tree using ggtree from a Newick formatted file.
}
\usage{
create_tree(tree = NULL, labels = FALSE)
}
\arguments{
  \item{tree}{A Newick (.nwk) constructed phylogenetic tree}
  \item{labels}{A boolean to determine whether to plot the labels on the tree structure}
}
