\name{get_sample_order_from_tree}
\alias{get_sample_order_from_tree}
\title{
Get sample order from tree data structure
}
\description{
Obtain a character vector of samples in the order of a tree structure.
}
\usage{
get_sample_order_from_tree(tree = NULL)
}

\arguments{
  \item{tree}{
A `tree` object from `read.tree` function
}
}
