\name{scale_y_tree}
\alias{scale_y_tree}

\title{
scale_y_tree
}
\description{
Default scales for continuous y values that correspond to the tree structure.
}
\usage{
scale_y_tree(expand = expansion(0, 0.6), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{expand}{A scale expansion vector (default: expansion(0, 0.6))}
  \item{\dots}{...}
}
